Extract state-holding `ConsumerFlux`
