+++ src/main/java/reactor/kafka/receiver/internals/AckMode.java
@@ -1,0 +1,8 @@
+package reactor.kafka.receiver.internals;
+
+enum AckMode {
+    AUTO_ACK,
+    MANUAL_ACK,
+    ATMOST_ONCE,
+    EXACTLY_ONCE,
+}
